Use go dep for dependency managment
Dependencies using go get are a bit of a mess, as some of our
dependencies "vendor" things (ie the src of their dependencies lives in
a vendor/ folder) and some don't. This partial vendoring causes some
libraries to be duplicated, and consequently their Init() function
called twice. golang.org/x/net/trace breaks (at runtime :() when this
happens.

Go dep vendors all of our dependencies, and ensures the vendor tree is
flat. This ensures we have a single copy of whatever library we need,
even if it's needed in several places.