NodeID: Use UUID instead of etcd member ID
The initial design hoped to use Maintenance alarms to monitor the health
of the cluster, and cleanup after dead nodes. The documentation for
these is a bit lacking however, and I could only get alarms when the
cluster actually went done - ie over half the nodes were gone.

This is replaced with a UUID & lease scheme:
- Node generate a UUID on startup
- Write that UUID to /node/ID using a lease
- Leader node watches /node/ID for deletions, which indicate a dead node