Task: Documentation, simplify newTask()
Document all the functions / methods properly, in a hopefully goish way.

Simplify newTask() to not enqeue tasks, just to initialize them. queue()
must be called afterwards (which api was doing anyways.. )

Document hacky Task version incrementing, seems to be the only way to do
it short of a new GET.