Runner: handle cancel(), requeue on error
Handle cancelled tasks in runner, by watch()'ing the tasks we're
currently running, and aborting if someone cancels (or modifies them in
another unexpected way)

Rework the runner to propagate errors upwards, and requeue the task on
error. Note that there is nothing preventing us from going in an
infinite loop of error, task requeuing and more errors.