Runner: process.wait() error handling botch
Unfortunatly process.Wait() returns an error if the process exits with a
none 0 exit code. The returned error is not a public type, so we
unfortunately can't check what kind of error it is.
The implementation shows that if a nil ProcessState is returned along
with an error, the process completed succesfully, just with a non zero
error code.